+++ Code/Engine/Foundation/Time/Implementation/Time_inl.h
@@ -26,6 +26,11 @@ inline ezTime::ezTime(const Seconds& seconds)
   operator=(seconds);
 }
 
+inline void ezTime::SetZero()
+{
+  m_fTime = 0.0;
+}
+
 inline void ezTime::operator=(const NanoSeconds& seconds)
 {
   m_fTime = seconds.m_fTime * 0.000000001;