+++ Documentation/WhatsNew.md
@@ -38,6 +38,7 @@ Milestone 3
   * Added optional *.natvis support for Visual Studio 2012. This enables much more comfortable inspection of basic types in the debugger.
   * The logging system has been improved. You can now log locally into another system (useful for tools) or you can replace the default implementation entirely.
   * Added ezClock for managing game time updates.
+  * Added ezStopwatch for measuring time.
 
 Milestone 2
 -----------